Split Inspector in 2 files
Just split in 2 responsabilities. Maybe configuration is not a good
name. Just a begin to learn more about the code.
